// frontend/bindings/collections/internal/model/models.ts
export class Attachment {
    "filename": string;
    "label"?: string;

    /** Creates a new Attachment instance. */
    constructor($$source: Partial<Attachment> = {}) {
        if (!("filename" in $$source)) {
            this["filename"] = "";
        }

        Object.assign(this, $$source);
    }

    /**
     * Creates a new Attachment instance from a string or object.
     */
    static createFrom($$source: any = {}): Attachment {
        let $$parsedSource = typeof $$source === 'string' ? JSON.parse($$source) : $$source;
        return new Attachment($$parsedSource as Partial<Attachment>);
    }
}

/**
 * Bookmark 书签实体
 */
export class Bookmark {
    "id": string;
    "url": string;
    "domain": string;
    "site_id": string;
    "title": string;
    "cover"?: string;
    "description"?: string;
    "tags": string[];
    "attachments": Attachment[];
    "status": string;
    "created_at": time$0.Time;
    "updated_at": time$0.Time;

    /** Creates a new Bookmark instance. */
    constructor($$source: Partial<Bookmark> = {}) {
        if (!("id" in $$source)) {
            this["id"] = "";
        }
        if (!("url" in $$source)) {
            this["url"] = "";
        }
        if (!("domain" in $$source)) {
            this["domain"] = "";
        }
        if (!("site_id" in $$source)) {
            this["site_id"] = "";
        }
        if (!("title" in $$source)) {
            this["title"] = "";
        }
        if (!("tags" in $$source)) {
            this["tags"] = [];
        }
        if (!("attachments" in $$source)) {
            this["attachments"] = [];
        }
        if (!("status" in $$source)) {
            this["status"] = "";
        }
        if (!("created_at" in $$source)) {
            this["created_at"] = null;
        }
        if (!("updated_at" in $$source)) {
            this["updated_at"] = null;
        }

        Object.assign(this, $$source);
    }

    /**
     * Creates a new Bookmark instance from a string or object.
     */
    static createFrom($$source: any = {}): Bookmark {
        const $$createField7_0 = $$createType0;
        const $$createField8_0 = $$createType2;
        let $$parsedSource = typeof $$source === 'string' ? JSON.parse($$source) : $$source;
        if ("tags" in $$parsedSource) {
            $$parsedSource["tags"] = $$createField7_0($$parsedSource["tags"]);
        }
        if ("attachments" in $$parsedSource) {
            $$parsedSource["attachments"] = $$createField8_0($$parsedSource["attachments"]);
        }
        return new Bookmark($$parsedSource as Partial<Bookmark>);
    }
}
// ...
const $$createType0 = $Create.Array($Create.Any);
const $$createType1 = Attachment.createFrom;
const $$createType2 = $Create.Array($$createType1);